Refatoração: criação das classes de serviço de Pokémon e Usuário

+++ src/main/java/com/menino/pokedexapi/domain/model/Usuario.java
@@ -13,7 +13,7 @@ import com.menino.pokedexapi.domain.dto.AlterarUsuarioDto;
 public class Usuario {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	Long id;
+	Integer id;
 	@NotBlank
 	String nome;
 	@NotBlank
@@ -25,7 +25,7 @@ public class Usuario {
 	
 	//Construtores
 	
-	public Usuario(Long id, AlterarUsuarioDto alterarUsuarioDto) {
+	public Usuario(Integer id, AlterarUsuarioDto alterarUsuarioDto) {
 		setId(id);
 		setNome(alterarUsuarioDto.getNome());
 		setEmail(alterarUsuarioDto.getEmail());
@@ -36,10 +36,10 @@ public class Usuario {
 	
 	//Getters e Setters
 	
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	public String getNome() {
